+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Revive Adserver will be documented in this file.
 
 - Fixed typo preventing Excel exports from working ([#1590](https://github.com/revive-adserver/revive-adserver/issues/1590))
 - Fixed UI bug in the *Hostname List* and *Registerable Domain List* delivery rules ([#1589](https://github.com/revive-adserver/revive-adserver/issues/1589))
+- Fixed issue with campaign start/end date when editing campaigns using non-english language ([#1588](https://github.com/revive-adserver/revive-adserver/issues/1588)) 
 
 
 ## [6.0.1] - 2025-10-24